fixing again

+++ gps_bot/gps.py
@@ -19,7 +19,7 @@ class InterfaceGPS:
             self.GPGGA_buffer = received_data.split("$GPGGA,",1)[1]
             self.NMEA_buff = (self.GPGGA_buffer.split(','))
             latitude, longitude = self.GPS_Info(self.NMEA_buff)
-            coordenates = {"latitude":int(latitude), "longitude" :int(longitude)}
+            coordenates = {"latitude":float(latitude), "longitude" :float(longitude)}
             return coordenates
 
     def GPS_Info(self, NMEA_buff):
